feat: refactor user dropdown component and rename file for clarity
